requirements.txt
streamlit==1.29.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
requests==2.31.0
statsmodels==0.14.1
scikit-learn==1.3.2 
statsmodels 
python-dotenv 